+++ ios/Classes/IosDeviceCheckPlugin.h
@@ -1,5 +1,4 @@
 #import <Flutter/Flutter.h>
-#import <DeviceCheck/DeviceCheck.h>
 
 @interface IosDeviceCheckPlugin : NSObject<FlutterPlugin>
 @end