+++ lib/ios_device_check.dart
@@ -1,18 +1,16 @@
-import 'dart:io';
-
 import 'package:flutter/services.dart';
 
 /// A Flutter plugin to use iOS DeviceCheck.
 /// 
 /// See: https://developer.apple.com/documentation/devicecheck
 class IosDeviceCheck {
-  static const _channel = MethodChannel('naokiokada.com/ios_device_check');
+  static const MethodChannel _channel =
+    MethodChannel('plugins.flutter.io/ios_device_check');
 
   IosDeviceCheck._();
 
-  /// A Boolean value indicating whether the DeviceCheck APIs are available on the current device.
-  static Future<bool> get isSupported async {
-    assert(Platform.isIOS);
+  /// A Boolean value indicating whether the DeviceCheck API is available on the current device.
+  static Future<bool> isSupported() async {
     return _channel.invokeMethod('isSupported');
   }
 
@@ -20,7 +18,6 @@ class IosDeviceCheck {
   /// 
   /// Converted to string from NSData from ios platform using Base64-encoding.
   static Future<String> generateToken() async {
-    assert(Platform.isIOS);
     return _channel.invokeMethod('generateToken');
   }
 }