+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## 1.1.0
+
+* upgrade flutter project files.
+* make `isSupported` to callable.
+
 ## 1.0.2
-* update doc
+
+* update doc.
 
 ## 1.0.1
+
 * internal improvement
 
 ## 1.0.0